# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Cenos API
  description: |
    API do ecossistema Cenos para gestão de agentes de campo (leituristas)
    das concessionárias de energia do Piauí (PI) e Maranhão (MA).

    ## Autenticação

    A API suporta 5 métodos de autenticação:

    1. **JWT Bearer** — Painel administrativo (`Authorization: Bearer <token>`)
    2. **Telegram TMA** — App do agente (`X-Telegram-Init-Data`)
    3. **Token Simples** — Consultas externas/BI (`?token=API_TOKEN`)

    Consulte a documentação detalhada em `docs/` para cada módulo.
  version: "2.0.0"
  contact:
    name: Cenos Suporte
    url: https://cenos.com.br
servers:
  - url: https://api.izi.tec.br
    description: Produção
  - url: http://localhost:3040
    description: Desenvolvimento
...
paths:
...
  /public/calendar:
    get:
      tags: [Público]
      summary: Calendário de eventos por estado/mês
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [pi, ma]
          description: Estado
        - name: month
          in: query
          schema:
            type: string
          description: Mês (YYYY-MM)
      responses:
        "200":
          description: Calendário
  /public/feriados:
    get:
      tags: [Público]
      summary: Feriados do estado
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [pi, ma]
          description: Estado
      responses:
        "200":
          description: Lista de feriados
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: 03/04/2026
...
  /public/generate_token:
    get:
      tags: [Público]
      summary: Gera token de acesso temporário
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          schema:
            type: string
          description: Telegram ID
      responses:
        "200":
          description: Token gerado
  /public/training/{id}:
    get:
      tags: [Público]
      summary: Visualização pública de treinamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Projeto de treinamento
        "404":
          description: Não encontrado
  /public/form/{id}:
    get:
      tags: [Público]
      summary: Estrutura pública de formulário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Estrutura do formulário
  /public/form/{id}/check:
    get:
      tags: [Público]
      summary: Verifica se respondente já respondeu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: respondentId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status da verificação
  /public/form/submit/{id}:
    post:
      tags: [Público]
      summary: Submete resposta de formulário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: object
                metadata:
                  type: object
      responses:
        "201":
          description: Resposta registrada
  /public/form/upload:
    post:
      tags: [Público]
      summary: Upload de arquivo para formulário
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: URL do arquivo
  /public/app_login:
    post:
      tags: [Público]
      summary: Login app nativo (matrícula + PIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matricula:
                  type: string
                pin:
                  type: string
      responses:
        "200":
          description: Token de acesso
...
components:
...
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Autenticação administrativa via JWT.
        Header: `Authorization: Bearer <token>`
        O token é obtido via `POST /admin/user/login`.
    TelegramAuth:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: |
        Autenticação do agente via Telegram Mini App.
        Header: `X-Telegram-Init-Data: <initData>`
        O initData é validado via Telegram TMA.
    ApiTokenAuth:
      type: apiKey
      in: query
      name: token
      description: |
        Token simples para consultas externas.
        Query param: `?token=API_TOKEN`
...